refactor(backend): group rollback ingress
